doucment_summarizer/pyproject.toml
[project]
name = "doucment_summarizer"
version = "0.1.0"
description = "document-summarizer using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-huggingface>=0.0.1",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "pypdf>=3.0.0",
    "pymupdf>=1.23.0",
    "unstructured>=0.10.0",
    "python-docx>=0.8.11",
    "pydantic>=2.0.0",
    "langchain-text-splitters>=0.0.1",
    "python-dotenv>=1.0.0"
]

[project.scripts]
doucment_summarizer = "doucment_summarizer.main:run"
run_crew = "doucment_summarizer.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
